requirements (2).txt
transformers==4.40.0
torch==2.2.0
torchvision==0.17.0
sentence-transformers==2.7.0
faiss-cpu==1.8.0
gradio==4.36.0
Pillow==10.3.0
pandas==2.2.0
numpy==1.26.4
accelerate==0.30.0
sentencepiece==0.2.0
protobuf==4.25.3
